Debugging

+++ src/util/ossimBatchTest.cpp
@@ -636,7 +636,7 @@ ossim_uint8 ossimBatchTest::processConfigList(const ossimKeywordlist& kwl)
          if (m_configFileName.isReadable())
          {
             overall_test_status |= execute();
-            cout<<"\n#### DEBUG 3 #### overall_test_status = "<<overall_test_status<<endl;
+            cout<<"\n#### DEBUG 3 #### overall_test_status = "<<(int)overall_test_status<<endl;
             ++processedIndexes;
          }
 
@@ -649,7 +649,7 @@ ossim_uint8 ossimBatchTest::processConfigList(const ossimKeywordlist& kwl)
       }
    }
 
-   cout<<"\n#### DEBUG 4 #### overall_test_status = "<<overall_test_status<<endl;
+   cout<<"\n#### DEBUG 4 #### overall_test_status = "<<(int)overall_test_status<<endl;
    return overall_test_status;
 }
 
@@ -787,13 +787,13 @@ ossim_uint8 ossimBatchTest::processTest(const ossimString& prefix, const ossimKe
    {
       ossimString prefixBase = prefix + "preprocess_command";
       testStatus |= processCommands( prefixBase, kwl, testName, false );
-      cout<<"\n#### DEBUG 1 #### testStatus = "<<testStatus<<endl;
+      cout<<"\n#### DEBUG 1 #### testStatus = "<<(int)testStatus<<endl;
    }
    if ( expectedFlag  && !(testStatus & TEST_ERROR))
    {
       ossimString prefixBase = prefix + "expected_results_command";
       testStatus |= processCommands( prefixBase, kwl, testName, true );
-      cout<<"\n#### DEBUG 2 #### testStatus = "<<testStatus<<endl;
+      cout<<"\n#### DEBUG 2 #### testStatus = "<<(int)testStatus<<endl;
    }
    if ( testFlag  && !(testStatus & TEST_ERROR))
    {
